requirements.txt
huggingface_hub
transformerS
accelerate
bitsandbytes
langchain
streamlit
python-dotenv
PyPDF2
langchain_community



-e .


# to download all the dependencies in the requirements.txt file run the following command in the terminal
  # pip install -r requirements.txt